Create LogStringHelper and move table string functions into it
